more feedback on web page enough stack for udp? now using float values for duty http and udp using now the range 0-255

+++ customer_app/custom_app/custom_app/pwm.c
@@ -25,35 +25,35 @@ void update_rgbw_duties(int red, int green, int blue, int white) {
     need_to_update_rgbw_duties = true;
 }
 
-void update_channel_duty(int channel_index, int duty) {
+void update_channel_duty(int channel_index, float duty) {
     if(channel_index <= -1 || channel_index >= 4) return;
-    if(duty < 0 || duty > 100) return;
+    if(duty < 0.f || duty > 100.f) return;
     current_channel = channel_index;
     current_channel_duty = duty;
     need_to_update_channel_duty = true;
 }
 
-void set_channel_duty(int channel_index, int duty) {
+void set_channel_duty(int channel_index, float duty) {
     if(channel_index <= -1 || channel_index >= 4) return;
-    if(duty < 0 || duty > 100) return;
+    if(duty < 0.f || duty > 100.f) return;
     bl_pwm_set_duty(channel_pin_map[channel_index][0], duty);
 }
 
-void set_color_duty(int color_to_channel_index, int duty) {
+void set_color_duty(int color_to_channel_index, float duty) {
     int index = color_to_channel_pin_map[color_to_channel_index];
     set_channel_duty(index, duty);
 }
 
-void set_red_duty(int duty) { set_color_duty(0, duty); }
-void set_green_duty(int duty) { set_color_duty(1, duty); }
-void set_blue_duty(int duty) { set_color_duty(2, duty); }
-void set_white_duty(int duty) { set_color_duty(3, duty); }
-void set_rgb_duty(int red_duty, int green_duty, int blue_duty) {
+void set_red_duty(float duty) { set_color_duty(0, duty); }
+void set_green_duty(float duty) { set_color_duty(1, duty); }
+void set_blue_duty(float duty) { set_color_duty(2, duty); }
+void set_white_duty(float duty) { set_color_duty(3, duty); }
+void set_rgb_duty(float red_duty, float green_duty, float blue_duty) {
     set_red_duty(red_duty);
     set_green_duty(green_duty);
     set_blue_duty(blue_duty);
 }
-void set_rgbw_duty(int red_duty, int green_duty, int blue_duty, int white_duty) {
+void set_rgbw_duty(float red_duty, float green_duty, float blue_duty, float white_duty) {
     set_rgb_duty(red_duty, green_duty, blue_duty);
     set_white_duty(white_duty);
 }
